decouple servicecomb kie which is introduced by example (#714)
