--- examples/router/servicecomb/client/conf/chassis.yaml
@@ -1,10 +1,0 @@
----
-#APPLICATION_ID: CSE optional
-cse:
-  service:
-    registry:
-      address: http://127.0.0.1:30100
-  config:
-    client:
-      serverUri: http://127.0.0.1:30110
-      type: servicecomb-kie
