--- examples/router/servicecomb/client/conf/microservice.yaml
@@ -1,3 +1,0 @@
----
-service_description:
-  name: orderService
