--- examples/router/servicecomb/serverV1/conf/chassis.yaml
@@ -1,13 +1,0 @@
----
-cse:
-  service:
-    registry:
-      address: http://127.0.0.1:30100 # If type is File then address will be the path of the file
-  protocols:
-    rest:
-      listenAddress: 127.0.0.1:5002
-      advertiseAddress: 127.0.0.1:5002
-  handler:
-    chain:
-      Provider:
-        default: tracing-provider,bizkeeper-provider
