--- examples/router/servicecomb/serverV2/conf/microservice.yaml
@@ -1,7 +1,0 @@
----
-#微服务的私有属性
-service_description:
-  name: paymentService
-  version: 2.0
-  instance_properties:
-    project: z
